Fixed issue where echo without option -n allowed creation of hex strings longer than 64 characters.

+++ project_up.sh
@@ -1,16 +1,25 @@
 #!/bin/sh
 
 project_up() {
-  cd ./security
+  local ENC_UTIL_IMAGE_NAME=the-supertask_wrapper
+
+  cd ./security 
   [ ! $? -eq 0 ] && echo -e "\033[7;31mThere Was a Problem with Command 1\033[0m" && return 1
 
-  # Build OpenSSL image.
-  docker build --no-cache \
-  --target wrap-enabled \
-  --rm --compress \
-  --file ./aes_wrap_enabled_openssl/Dockerfile \
-  .
-  [ ! $? -eq 0 ] && echo -e "\033[7;31mThere Was a Problem with Command 2\033[0m" && return 1
+  # Only Build AES Wrap-Enabled OpenSSL if necessary.
+  if [ -z "$(docker images | grep -o ${ENC_UTIL_IMAGE_NAME})" ]; then
+    echo -e "\033[7;31m\n                                                  \nImage ${ENC_UTIL_IMAGE_NAME} Needs to be Built :( \n                                                  \033[0m"
+    # Build OpenSSL image.
+    docker build --no-cache \
+    --tag the-supertask_wrapper:latest \
+    --target wrap-enabled \
+    --rm --compress \
+    --file ./aes_wrap_enabled_openssl/Dockerfile \
+    .
+    [ ! $? -eq 0 ] && echo -e "\033[7;31mThere Was a Problem with Command 2\033[0m" && return 1
+  else
+    echo -e "\033[7;32m\n                                              \n Image ${ENC_UTIL_IMAGE_NAME} Already Built :) \n                                              \033[0m"
+  fi
 
   # Run instance of image.
   docker run -d --entrypoint "/bin/sh" -it --name enc_util \
@@ -20,7 +29,7 @@ project_up() {
   --mount type=volume,src=persist_pgp,dst=/pgp/ \
   --mount type=volume,src=persist_tls,dst=/tls/ \
   --rm=false \
-  the-supertask_wrapper
+  ${ENC_UTIL_IMAGE_NAME}
   [ ! $? -eq 0 ] && echo -e "\033[7;31mThere Was a Problem with Command 3\033[0m" && return 1
 
   # Copy files into  container.
@@ -30,12 +39,16 @@ project_up() {
   [ ! $? -eq 0 ] && echo -e "\033[7;31mThere Was a Problem with Command 4\033[0m" && return 1
 
   # Execute commands inside container.
-  docker exec -it enc_util \
+  docker exec -it \
+  -e OPENSSL_V111=/root/local/bin/openssl.sh \
+  enc_util \
   /bin/sh -c '. /etc/profile && \
   run_install --verbose /etc/profile.d/install_pgp.sh'
   [ ! $? -eq 0 ] && echo -e "\033[7;31mThere Was a Problem with Command 5\033[0m" && return 1
 
-  docker exec -it enc_util \
+  docker exec -it \
+  -e OPENSSL_V111=/root/local/bin/openssl.sh \
+  enc_util \
   /bin/sh -c '. /etc/profile && \
   run_install --verbose /etc/profile.d/install_tls.sh'
   [ ! $? -eq 0 ] && echo -e "\033[7;31mThere Was a Problem with Command 6\033[0m" && return 1
